+++ samples/qa-voice-control/src/views/intent/IntentDelete.js
@@ -19,7 +19,6 @@ class IntentDelete extends React.Component {
 
 	showResult = (msg) => {
 		this.setState({result: msg});
-		setTimeout(() => (this.setState({result: ''})), 1500);
 	};
 
 	handleVoice = (e) => {