+++ RangePicker/RangePicker.js
@@ -265,7 +265,7 @@ const RangePickerBase = kind({
 	render: ({label, value, voiceLabel, ...rest}) => {
 		delete rest.padded;
 		return (
-			<Picker {...rest} css={css} data-webos-voice-labels-ext={voiceLabel} index={0} value={value} reverse={false}>
+			<Picker {...rest} css={css} data-webos-voice-labels-ext={voiceLabel} index={0} reverse={false} type="number" value={value}>
 				<PickerItem key={value} marqueeDisabled style={{direction: 'ltr'}}>{label}</PickerItem>
 			</Picker>
 		);