+++ TimePicker/TimePickerBase.js
@@ -317,6 +317,7 @@ const TimePickerBase = kind({
 					const isLeft = rtl && picker === 'a' || isFirst && !rtl;
 					// minute will always be the right-most control in RTL, regardless of the provided order
 					const isRight = rtl && picker === 'm' || isLast && !rtl;
+					const isLastElement = rtl ? isLeft : isLast;
 
 					switch (picker) {
 						case 'h':
@@ -327,9 +328,10 @@ const TimePickerBase = kind({
 										accessibilityHint={hourAccessibilityHint}
 										aria-label={hourAriaLabel}
 										className={css.hourPicker}
-										disabled={disabled}
+										data-last-element={isLastElement}
 										data-webos-voice-disabled={voiceDisabled}
 										data-webos-voice-group-label={hourAccessibilityHint}
+										disabled={disabled}
 										hasMeridiem={hasMeridiem}
 										onChange={onChangeHour}
 										onSpotlightDisappear={onSpotlightDisappear}
@@ -349,9 +351,10 @@ const TimePickerBase = kind({
 									accessibilityHint={minuteAccessibilityHint}
 									aria-label={minuteAriaLabel}
 									className={css.minutePicker}
-									disabled={disabled}
+									data-last-element={isLastElement}
 									data-webos-voice-disabled={voiceDisabled}
 									data-webos-voice-group-label={minuteAccessibilityHint}
+									disabled={disabled}
 									key="minute-picker"
 									max={59}
 									min={0}
@@ -372,9 +375,10 @@ const TimePickerBase = kind({
 									aria-label={meridiemAriaLabel}
 									aria-valuetext={meridiems ? meridiems[meridiem] : null}
 									className={css.meridiemPicker}
-									disabled={disabled}
+									data-last-element={isLastElement}
 									data-webos-voice-disabled={voiceDisabled}
 									data-webos-voice-group-label={meridiemLabel}
+									disabled={disabled}
 									key="meridiem-picker"
 									onChange={onChangeMeridiem}
 									onSpotlightDisappear={onSpotlightDisappear}
@@ -397,6 +401,15 @@ const TimePickerBase = kind({
 		);
 	}
 });
+
+/**
+ * Called when `Enter` key down on the last picker
+ *
+ * @name onComplete
+ * @memberof sandstone/TimePicker.TimePickerBase.prototype
+ * @type {Function}
+ * @public
+ */
 
 export default TimePickerBase;
 export {TimePickerBase};