+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 - `sandstone/DatePicker` and `sandstone/TimePicker` to show arrows when normal
 
+### Changed
+
+- `sandstone/TooltipDecorator` not to read out audio guidance when showing
+
 ## [1.2.1] - 2020-09-03
 
 ### Changed