+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
-## [unreleased]
+## [1.3.0] - 2020-09-14
 
 ### Added
 
@@ -16,9 +16,9 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Fixed
 
-- `sandstone/Scroller` to not restrict focus movement with 5-way directional keys when `focusableScrollbar` prop is `byEnter` and there is no scrollbar
-- `sandstone/DatePicker` and `sandstone/TimePicker` to focus next picker with enter key
+- `sandstone/DatePicker` and `sandstone/TimePicker` to focus the next picker with enter key
 - `sandstone/DatePicker` and `sandstone/TimePicker` to show arrows when normal
+- `sandstone/Scroller` to not restrict focus movement with 5-way directional keys when `focusableScrollbar` prop is `byEnter` and there is no scrollbar
 
 ## [1.2.1] - 2020-09-03